+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Use ggwsm in pipelines
+- Resolve dependency cycles that close over a dev dependency
 
 ## 4.1.0 - 2026-08-13